# openapi.yaml
openapi: 3.0.3
info:
  title: MOSS API
  description: REST API for the MOSS metadata server.
  version: 0.2.0
servers:
  - url: http://localhost:8080
    description: Local dev
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
paths:
...
  /g/{path}:
    x-jaxrs-base-path: "/g"
    get:
      tags: [metadata]
      operationId: getMetadataResource
      x-jaxrs-path: "/{path:.*}"
      x-required-permission: read-metadata
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/turtle:
              schema:
                type: string
            application/ld+json:
              schema:
                type: string
...